+++ BetterItemHandling/Plugin.cs
@@ -6,7 +6,9 @@ using BepInEx;
 using BepInEx.Configuration;
 using BepInEx.Logging;
 using BetterItemHandling.Data;
+using GameNetcodeStuff;
 using HarmonyLib;
+using System.Reflection;
 
 namespace BetterItemHandling
 {
@@ -18,13 +20,14 @@ namespace BetterItemHandling
     {
         private const string ModGUID = "Yan01h.BetterItemHandling";
         private const string ModName = "BetterItemHandling";
-        private const string ModVersion = "1.0.0";
+        private const string ModVersion = "1.1.0";
 
         // Main logger used by this plugin
         public static ManualLogSource Log {  get; private set; }
 
         // Config entries
         public static ConfigEntry<bool> ConfigAllowDropAllScrap;
+        public static ConfigEntry<bool> ConfigAllowPlaceAllScrapOnDesk;
 
         private readonly Harmony _harmony = new Harmony(ModGUID);
 
@@ -38,6 +41,10 @@ namespace BetterItemHandling
             Log.LogInfo($"Loading {ModName} (v{ModVersion})");
 
             ConfigAllowDropAllScrap = Config.Bind("General", "AllowDropAllScrap", true, "Allows dropping all scrap when pressing your drop key on an empty slot");
+            ConfigAllowPlaceAllScrapOnDesk = Config.Bind("General", "AllowPlaceAllScrapOnDesk", true, "Allows to quickly place all scrap in your inventory on the deposit desk when pressing your sprint key");
+
+            PlayerControllerData.SwitchToItemSlot = typeof(PlayerControllerB)
+                .GetMethod("SwitchToItemSlot", BindingFlags.Instance | BindingFlags.NonPublic);
 
             _harmony.PatchAll();