+++ thunderstore/CHANGELOG.md
@@ -5,6 +5,16 @@ All notible changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),  
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0]
+
+### Added
+
+- Added the ability to instantly place all scrap on the deposit desk **(Can be turned off in the config)**.
+
+### Fixed
+
+- Fixed a bug that allowed scrolling through the inventory while spamming g with a two handed item.
+
 ## [1.0.0]
 
-- Initial release
+- Initial release.